modules/location/requirements.txt
Flask
Flask-Cors
Flask-Restx
Flask-Script
Flask-Migrate
psycopg2-binary
SQLAlchemy
kafka-python
geoalchemy2
shapely